move leaf field validation into field validation
